+++ qiskit/providers/models/qobj/__init__.py
@@ -1,10 +1,13 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2018, IBM.
+# Copyright 2019, IBM.
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
 """Models for Qobj and its related components."""
 
+from .models import (QobjConditional, QobjConfig, QobjExperiment,
+                     QobjExperimentConfig, QobjExperimentHeader, QobjHeader,
+                     QobjInstruction)
 from .qobj import Qobj
